Relocate CountyContext unit tests from test_partisan.py to test_county_context.py
ideals_for_eguia retry/cache tests, _ensure_county_data tests, and the
_populate_county_data materialized-view guard test were pure CountyContext
unit tests that had accumulated in test_partisan.py simply because
CountyContext originally existed to serve the Eguia metric. They test
CountyContext itself, not partisan scoring, so they belong alongside the
other CountyContext tests in test_county_context.py. Genuinely eguia-metric
integration tests (calling eguia_county() directly) stay in test_partisan.py.
