+++ backend/tests/test_county_context.py
@@ -1,4 +1,5 @@
-"""Tests for CountyContext's forced-minimum precompute (app.evaluation.context).
+"""Tests for CountyContext (app.evaluation.context) — the singleton behind both
+the Eguia metric's county ideals and the county-pieces forced-minimum precompute.
 
 component_populations on evaluation.county_demographics is document-independent:
 each county's own connected components (VTD/parent-unit adjacency, ignoring any
@@ -6,9 +7,17 @@ document/assignment) and the population of each. Used to compute the forced-
 minimum split count as sum(ceil(p/ideal_pop) for p in component_populations),
 a tighter bound than ceil(total_pop/ideal_pop) for counties that are themselves
 already several disconnected land pieces (islands, exclaves).
+
+The ideals_for_eguia / _ensure_county_data / _populate_county_data unit tests
+below were relocated from test_partisan.py, where they'd accumulated simply
+because CountyContext originally existed to serve the Eguia metric — they test
+CountyContext itself, not partisan scoring, so they belong here. Genuinely
+eguia-metric integration tests (calling eguia_county() and asserting on its
+behavior) stay in test_partisan.py.
 """
 
 from datetime import datetime
+from unittest.mock import MagicMock
 
 import fastapi
 import pytest
@@ -20,11 +29,15 @@ from sqlmodel import Session, select
 import app.evaluation.context as context_module
 from app.evaluation.context import (
     COUNTY_CONTEXT,
+    CountyContext,
     DocumentEvaluationContext,
+    GerrydbTableName,
     component_populations_for_nodes,
 )
 from app.evaluation.models import CountyDemographics
 from app.evaluation.splits import county_pieces
+
+_STUB_TABLE = GerrydbTableName("test_table")
 
 
 def test_component_populations_for_nodes_single_component():
@@ -55,6 +68,107 @@ def test_component_populations_for_nodes_ignores_unlisted_nodes():
     result = component_populations_for_nodes(G, populations)
     # "a" and "c" are only connected through "b", which isn't in the node set.
     assert sorted(result) == [100, 300]
+
+
+# ── ideals_for_eguia: retry/cache behavior (relocated from test_partisan.py) ──
+
+
+def test_ideals_for_eguia_retries_then_gives_up():
+    """Failures are retried up to MAX_LOAD_ATTEMPTS times; subsequent calls
+    raise immediately without hitting the DB."""
+    table = GerrydbTableName(_STUB_TABLE)
+    singleton = CountyContext()
+    singleton._ensure_county_data = MagicMock()
+    mock_compute = MagicMock(side_effect=ValueError("no data"))
+    singleton._compute_ideal = mock_compute
+
+    for attempt in range(1, CountyContext.MAX_LOAD_ATTEMPTS + 1):
+        with pytest.raises(ValueError):
+            singleton.ideals_for_eguia(table, MagicMock())
+        assert mock_compute.call_count == attempt
+
+    # After exhausting attempts, raises without additional DB work.
+    with pytest.raises(ValueError, match="failed to load after"):
+        singleton.ideals_for_eguia(table, MagicMock())
+    assert mock_compute.call_count == CountyContext.MAX_LOAD_ATTEMPTS
+
+
+def test_ideals_for_eguia_recovers_after_transient_failure():
+    """A transient failure on one attempt doesn't block a later successful
+    attempt; once recovered, the result is permanently cached and _compute_ideal
+    is never called again."""
+    table = GerrydbTableName(_STUB_TABLE)
+    good_ideals = {"pres_2020_dem": 0.6, "pres_2020_rep": 0.4}
+    singleton = CountyContext()
+    singleton._ensure_county_data = MagicMock()
+    mock_compute = MagicMock(side_effect=[ValueError("transient"), good_ideals])
+    singleton._compute_ideal = mock_compute
+
+    with pytest.raises(ValueError):
+        singleton.ideals_for_eguia(table, MagicMock())
+    assert singleton._attempts[table] == 1
+
+    assert singleton.ideals_for_eguia(table, MagicMock()) == good_ideals
+    assert singleton._cache[table] == good_ideals
+
+    # All further calls hit the cache — _compute_ideal is not called again.
+    assert singleton.ideals_for_eguia(table, MagicMock()) == good_ideals
+    assert mock_compute.call_count == 2
+
+
+# ── _ensure_county_data (relocated from test_partisan.py) ─────────────────────
+
+
+def test_ensure_county_data_calls_populate_when_no_valid_rows():
+    """_ensure_county_data triggers _populate_county_data when no row with
+    non-null total_pop exists (covers both the no-rows and null-total_pop cases)."""
+    table = GerrydbTableName(_STUB_TABLE)
+    singleton = CountyContext()
+    mock_populate = MagicMock()
+    singleton._populate_county_data = mock_populate
+
+    mock_session = MagicMock()
+    mock_session.exec.return_value.first.return_value = None
+
+    singleton._ensure_county_data(table, mock_session)
+
+    mock_populate.assert_called_once_with(table, mock_session)
+
+
+def test_ensure_county_data_skips_populate_when_valid_rows_exist():
+    """_ensure_county_data is a no-op when a row with non-null total_pop exists."""
+    table = GerrydbTableName(_STUB_TABLE)
+    singleton = CountyContext()
+    mock_populate = MagicMock()
+    singleton._populate_county_data = mock_populate
+
+    mock_session = MagicMock()
+    mock_session.exec.return_value.first.return_value = MagicMock()
+
+    singleton._ensure_county_data(table, mock_session)
+
+    mock_populate.assert_not_called()
+
+
+# ── _populate_county_data: materialized-view guard (relocated from test_partisan.py) ──
+#
+# Retesting for a past bug: eguia county aggregation must use parent_layer (VTD
+# base table), not the shatterable UNION ALL materialized view.
+
+
+def test_populate_county_data_rejects_materialized_view(
+    session, gerrydb_ks_ellis_geos_view
+):
+    """_populate_county_data must raise for materialized views.
+
+    The shatterable gerrydb view (ks_ellis_geos) is a UNION ALL of VTD and
+    block rows. Inserting from it would double-count every county.
+    The pg_class.relkind guard must raise rather than silently skip.
+    """
+    shatterable_view = GerrydbTableName("ks_ellis_geos")
+
+    with pytest.raises(ValueError, match="plain table"):
+        CountyContext()._populate_county_data(shatterable_view, session)
 
 
 def test_simple_geos_county_demographics_component_populations(
